refactor(pendle-v2): move token fetchers to common folder
